+++ src/src/cn/limc/androidcharts/entity/IHasDate.java
@@ -41,12 +41,12 @@ public interface IHasDate {
 	/**
 	 * @return the date
 	 */
-	int getDate();
+	long getDate();
 
 	/**
 	 * @param date
 	 *            the date to set
 	 */
-	void setDate(int date);
+	void setDate(long date);
 
 }
